+++ pya/__init__.py
@@ -8,9 +8,8 @@ from .Aserver import Aserver
 from .Arecorder import Arecorder
 from .Ugen import Ugen
 from .version import __version__
-from .helper import ampdb, dbamp, cpsmidi, midicps, linlin, clip
-from .helper import audio_from_file, buf_to_float, spectrum, audio_from_file
-from .helper import normalize, record, device_info
+from .helper import *
+from .backend import *
 
 
 def startup(**kwargs):
@@ -19,11 +18,3 @@ def startup(**kwargs):
 
 def shutdown(**kwargs):
     Aserver.shutdown_default_server(**kwargs)
-
-
-__all__ = ['Asig', 'Aspec', 'Astft', 'Aserver', 'Arecorder', 'Ugen',
-           'startup', 'shutdown']
-
-__all__ += ['ampdb', 'dbamp', 'cpsmidi', 'midicps', 'linlin', 'clip', 
-            'audio_from_file', 'buf_to_float', 'spectrum', 
-            'audio_from_file', 'normalize', 'record', 'device_info']